fix lr 5

+++ src/main/java/ua/com/kisit/course2026np/repository/CrewMemberRepository.java
@@ -5,15 +5,16 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import ua.com.kisit.course2026np.entity.CrewMember;
+import ua.com.kisit.course2026np.entity.CrewRole;
 
 import java.util.List;
 
 @Repository
-public interface CrewMemberRepository extends JpaRepository<CrewMember, Long> {
-    List<CrewMember> findByCrewId(Long crewId);
-    List<CrewMember> findByCrewIdAndRole(Long crewId, String role);
-    long countByCrewIdAndRole(Long crewId, String role);
+public interface CrewMemberRepository extends JpaRepository<CrewMember, Integer> {
+    List<CrewMember> findByCrewId(Integer crewId);
+    List<CrewMember> findByCrewIdAndRole(Integer crewId, String role);
+    long countByCrewIdAndRole(Integer crewId, String role);
     @Query("SELECT cm FROM CrewMember cm WHERE cm.crew.flight.id = :flightId")
-    List<CrewMember> findByFlightId(@Param("flightId") Long flightId);
+    List<CrewMember> findByFlightId(@Param("flightId") Integer flightId);
     List<CrewMember> findByLastNameContainingIgnoreCase(String lastName);
 }
