+++ src/main/java/ua/com/kisit/course2026np/repository/CrewRepository.java
@@ -7,14 +7,8 @@ import ua.com.kisit.course2026np.entity.Crew;
 import java.util.Optional;
 
 @Repository
-public interface CrewRepository extends JpaRepository<Crew, Long> {
-
-    // Бригада для конкретного рейсу (flight_id_UNIQUE)
-    Optional<Crew> findByFlightId(Long flightId);
-
-    // Чи існує вже бригада для цього рейсу
-    boolean existsByFlightId(Long flightId);
-
-    // Видалити бригаду разом з усіма членами (orphanRemoval = true)
-    void deleteByFlightId(Long flightId);
+public interface CrewRepository extends JpaRepository<Crew, Integer> {
+    Optional<Crew> findByFlightId(Integer flightId);
+    boolean existsByFlightId(Integer flightId);
+    void deleteByFlightId(Integer flightId);
 }
