+++ src/main/java/ua/com/kisit/course2026np/repository/FlightRepository.java
@@ -5,13 +5,14 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import ua.com.kisit.course2026np.entity.Flight;
+import ua.com.kisit.course2026np.entity.FlightStatus;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface FlightRepository extends JpaRepository<Flight, Long> {
+public interface FlightRepository extends JpaRepository<Flight, Integer> {
 
     // Пошук рейсу за унікальним номером
     Optional<Flight> findByFlightNumber(String flightNumber);
@@ -20,7 +21,7 @@ public interface FlightRepository extends JpaRepository<Flight, Long> {
     List<Flight> findByStatus(String status);
 
     // Рейси за адміністратором (індекс idx_user_id)
-    List<Flight> findByUserId(Long userId);
+    List<Flight> findByUserId(Integer userId);
 
     // Рейси за маршрутом
     List<Flight> findByDepartureCityAndArrivalCity(
